series of bugfixes on pyramids and bigrams
