+++ sktime/transformers/series_as_features/dictionary_based/_sax.py
@@ -171,39 +171,3 @@ class SAX(BaseSeriesAsFeaturesTransformer):
             10: [-1.28, -0.84, -0.52, -0.25, 0.0, 0.25, 0.52, 0.84, 1.28,
                  sys.float_info.max]
         }[self.alphabet_size]
-
-
-class _BitWord(object):
-    # Used to represent a word for dictionary based classifiers such as BOSS
-    # an BOP.
-    # Can currently only handle an alphabet size of <= 4 and word length of
-    # <= 16.
-    # Current literature shows little reason to go beyond this, but the
-    # class will need changes/expansions
-    # if this is needed.
-    # TODO a shift of 2 is only correct for alphabet size 4, log2(4)=2
-
-    @staticmethod
-    def create_bigram_word(word, other_word, length):
-        return (word << (2 * length)) | other_word
-
-    @classmethod
-    def shorten_word(cls, word, amount):
-        # shorten a word by set amount of letters
-        return cls.right_shift(word, amount * 2)
-
-    @classmethod
-    def word_list(cls, word, length):
-        # list of input integers to obtain current word
-        word_list = []
-        shift = 32 - (length * 2)
-
-        for i in range(length - 1, -1, -1):
-            word_list.append(cls.right_shift(word << shift, 32 - 2))
-            shift += 2
-
-        return word_list
-
-    @staticmethod
-    def right_shift(left, right):
-        return (left % 0x100000000) >> right
